refactor service checks a bit
